moving the process of saving edits
